==> Python/00_data-ingest-CMAP.ipynb <==
# Loading Functions

import pandas as pd
import numpy as np


## Creating Real Time data set
### Datetime is in UTC

realtime_path = '/Users/cristianswift/Desktop/Spring-Quarter-2022-2023/SeniorThesis/data/original/SeaFlow_realtime_TN413.csv'

#changing column naes so that the date is just called data
realtime_cols = ['time',
 'lat',
 'lon',
 'population',
 'abundance_cells_per_microliter',
 'diameter_micrometer']

#reading in the csv to a pandas df
realtime = pd.read_csv(realtime_path, names=realtime_cols)
# getting rid of the first row since it has the column names repeated
realtime = realtime.tail(-1)
realtime

## Averaging data over 10 minute resolution so that SeaFlow and Underway CTD data match

realtime['abundance_cells_per_microliter'] = pd.to_numeric(realtime['abundance_cells_per_microliter'])

realtime = realtime.dropna(subset=['time','lat', 'lon']).reset_index().drop('index',axis=1)
realtime['time'] = realtime['time'].str[:-1]
realtime['lat'] = realtime['lat'].astype(float)
realtime['lon'] = realtime['lon'].astype(float)
realtime['abundance_cells_per_microliter'] = realtime['abundance_cells_per_microliter'].astype(float)
realtime['diameter_micrometer'] = realtime['diameter_micrometer'].astype(float)
realtime['depth'] = 0

print(realtime.dtypes)
realtime.head(4)

#averaging over every 10 min
realtime['time'] = pd.to_datetime(realtime['time'])
#this drops the phytoplankton population categorical column, need to recalulate
realtime_avg = (realtime
                .groupby('population')
                .resample('10min', on='time')
                .mean(numeric_only=False)
                .reset_index()
               )
realtime_avg
# making time a string object again
realtime_avg['time'] = realtime_avg['time'].astype(str)
realtime_avg['time'] = realtime_avg['time'].str.replace(' ', 'T')

# Using SimonCMAP to gather additional features

#### First installing and importing pycmap 

# !pip install pycmap
import pycmap

### Prepping realtime data for colocalization using Simon's CMAP

#### Setting API

api = pycmap.API(token='<6e1eb1d3-d364-4dfb-9121-8c23369dbbbe>')

targets = {
        
        # BioGeoChemical Numerical Near-Real-Time Model
        "tblPisces_Forecast_cl1": {
                          "variables": ["NO3", "PO4", "Fe", "Si", "chl", "nppv", "temp", "salin"],
                          "tolerances": [4, 0.5, 0.5, 5]
                         }
        }


source = realtime_avg

realtime_cmap = pycmap.Sample(
              source=source, 
              targets=targets, 
              replaceWithMonthlyClimatolog=False
             )


#saving as a CSV file
realtime_cmap.to_csv('/Users/cristianswift/Desktop/Spring-Quarter-2022-2023/SeniorThesis/data/modified/realtime_cmap.csv', index=False)


## The given Covariate dataset has a lot of NaNs values for nutrient information retrieved from CMAP, we will try to use Two Pisces Climatological models to fill in the gaps.

covari_path = '/Users/cristianswift/Desktop/Spring-Quarter-2022-2023/SeniorThesis/data/original/SeaFlow_covariates.csv'
covari = pd.read_csv(covari_path, usecols=lambda column: column != 'Unnamed: 0')
#keeping only relavent columns
covari = covari[['date_UTC', 'PopulationName', 'Latitude_decimalDegree', 'Longitude_decimalDegree',
        'Biomass_pgC_per_L', 'Salinity_psu', 'Temperature_degC']]


# rename for easier use and making time a string object to be manipulated
covari.rename(columns={
    'date_UTC': 'time', 'Latitude_decimalDegree': 'lat', 'Longitude_decimalDegree': 'lon'}, inplace=True)
covari['time'] = covari['time'].astype(str)
#making date time format CMAP appropriate
covari['time'] = covari['time'].str.replace('Z', '')
covari

targets = {
        
        # BioGeoChemical Numerical Near-Real-Time Model
        "tblPisces_NRT": {
                          "variables": ["NO3", "PO4", "Fe", "Si", "chl"],
                          "tolerances": [4, 0.5, 0.5, 5]
                         },
    
    "tblPisces_Forecast": {
                          "variables": ["NO3", "PO4", "Fe", "Si", "chl"],
                          "tolerances": [4, 0.5, 0.5, 5]
        }
}
import pycmap
api = pycmap.API(token='<6e1eb1d3-d364-4dfb-9121-8c23369dbbbe>')

source = covari
covari_cmap = pycmap.Sample(
              source=source, 
              targets=targets, 
              replaceWithMonthlyClimatolog=False
             )


Saving as a CSV to be accesible later.

covari_cmap.to_csv('/Users/cristianswift/Desktop/Spring-Quarter-2022-2023/SeniorThesis/data/modified/covari_cmap.csv', index=False)
