==> Python/02_data-cleaning.ipynb <==
## This notebook we will be cleaning the data of any NaN values, and getting it ready for machine learning applications.

#Running data ingest python notebook

%run 01_data-ingest-cmap.ipynb

### Need to adjust date, lat and lon columns so that they are in the appropriate formates

print(pd.DataFrame(covari))
print('\n')
print(covari.dtypes)
print('\n')

### Need to ajust columns that contain numbers to floats

def ChangeObjectTypes(df):
    for column in df:
        if column == 'date':
            #changing to datetime
            df[column] = pd.to_datetime(df[column], errors='coerce')
            
        elif column == 'PopulationName' or column == 'cruisename':
            #changing to string
            df[column] = df[column].astype(str)
            
        else:
            #changing to numeric type
            df[column] = pd.to_numeric(df[column], errors='coerce')
    return df
        

ChangeObjectTypes(covari)

### Now all of the objects are in the right format

print(pd.DataFrame(covari))
print('\n')
print(covari.dtypes)
print('\n')

import matplotlib.pyplot as plt
import pandas as pd

def plot_nans(df):
    """
    This function takes a DataFrame and determines if there are NaN values. If there are NaN values, it returns a
    DataFrame that has the ratio of NaN values for each column, along with a bar graph showing the ratio of NaNs.
    """
    nan_counts = df.isnull().sum()
    total_rows = len(df)
    
    if nan_counts.sum() != 0:
        nan_ratio = nan_counts / total_rows * 100
        nan_df = pd.DataFrame({'Variables': nan_ratio.index, 'Ratio of Data Missing': nan_ratio.values})
        nan_df = nan_df[nan_df['Ratio of Data Missing'] > 0].sort_values(by='Ratio of Data Missing', ascending=False)
        
        # Plotting the NaN values
        fig, ax = plt.subplots(figsize=(10, 6))
        ax.bar(nan_df['Variables'], nan_df['Ratio of Data Missing'])
        plt.xticks(rotation=45, ha='right')
        ax.set_title('Missing Data of Covariate DataFrame')
        ax.set_ylabel('NaNs (%)')
        ax.grid(axis='y')
        plt.tight_layout()
        plt.show()
        
        return nan_df
    else:
        print("No NaNs found. Yay!")
        return None


plot_nans(covari)

### Creating seperate dataframes of covar that have Nans and that do not

covari_nans = covari
covari = covari.dropna()

covari

##  Combining picoeuk and nanoeuk to one Population name (non_pro_syn)

covari['PopulationName'] = covari['PopulationName'].replace(
    {'nanoeukaryotes (2-5µm)': 'non_pro_syn', 'picoeukaryotes (< 2µm)': 'non_pro_syn'}
)
        

tn413

## Calculating Biomaass

### Seaflow uses 
1. Cell abundance
2. Cell Diameter

To Calulate Biomass we first need to take the diameter of the plankton and find the volume using:
d = diameter
r = radius
QC = Carbon Quotas


1. $\frac{Diameter}{2}\ = r$

2. $\frac{4}{3}\pi r^3 = biovolume$

3. *Qc = 0.261 * Biovolume <sup>0.86</sup>*

4. $biomass = Qc * abundance$

#first calulating volumn of a sphere given plankton diamter
import math
plankton_volume = (4/3) * math.pi * (realtime_cmap['diameter_micrometer']/2)**3
#calulating carbon quota
carbon_quota = 0.261 * plankton_volume**0.86
# multipling the result by abundance to get total biomass per measurement (multipling by 1000 to get pgC per L)
realtime_cmap['biomass'] = realtime_cmap['abundance_cells_per_microliter'] * carbon_quota

## Combinging some of the senior cruise data to covariate

import pandas as pd

# Randomly sample 30% of rows from tn413 DataFrame
tn413['cruisename'] = 'TN413'
tn413.drop(['cell_abundance', 'cell_diameter'], axis=1, inplace=True)


stratum_column = 'lat'
sample_proportion = 0.3

# Calculate stratum sizes
stratum_sizes = (tn413.groupby(stratum_column).size() * sample_proportion).astype(int)

# Sample each group
sampled_groups = []
for group_name, group_data in tn413.groupby(stratum_column):
    sampled_group = group_data.sample(n=stratum_sizes[group_name])
    sampled_groups.append(sampled_group)

# concatenate sampled groups
stratified_sample = pd.concat(sampled_groups)

# Concatenate the sampled rows to covari DataFrame
model_data = pd.concat([covari, sampled_rows], ignore_index=True)

model_data

#renaming these
model_data['PopulationName'] = model_data['PopulationName'].replace(
    {'nanoeukaryotes (2-5µm)': 'non_pro_syn', 'picoeukaryotes (< 2µm)': 'non_pro_syn',
    'picoeuk': 'non_pro_syn', 'synecho': 'non_pro_syn',
     'synecho': 'Synechococcus', 'prochloro':'Prochlorococcus'
    }
)
model_data.to_csv('/Users/cristianswift/Desktop/Spring-Quarter-2022-2023/SeniorThesis/data/model_data.csv', index=False)


model_data

# # convert the 'time' column to datetime format
# realtime['time'] = realtime['time'].str[:-1]
# realtime['time'] = pd.to_datetime(realtime['time'])

# # set the 'time' column as the index
# realtime.set_index('time', inplace=True)

# # resample the dataframe by hour and take the mean of each group
# realtime = realtime.resample('H').mean()

# # print the result
# realtime
